Smith: Add upstream mutator into the mix
It seems to help slightly, but it's hard to tell objectively because by
now my corpus is getting saturated anyway. It does now print the name of
the mutator in the mutation string, I wonder how it does that, there is
no API for it. Does it internally append something to a global? I want
to mutate that global too, would be useful to see what strategies are
effective so I can bias towards them.