Add custom WyRand implementation
It's not that there is something wrong with tinyrand, but it turns out,
WyRand is such a tiny function, I can implement it myself and avoid a
dependency. Do keep tinyrand for one more commit so I can write a test
to use it as a reference implementation. It passes.