Smith: Synthesize function types
This was one part that was still missing, which I suspect is why it was
failing to discover one uncovered line in the TypeDiff reporting that I
was hunting for. After a few minutes of running this it sitll hasn't
found it, but it is making steady progress and discovering lots of new
branches.