Smith: Generate short idents with a short instruction
I noticed the fuzzer was always using "std" and other builtins as
variable names, never using its own even though there is an instruction
for it. I suspect it's because of the REDUCE step in libFuzzer, using a
builtin saves one byte on the fuzz input size.

So add an equally short encoding for short identifiers, that is also
lexicographically smaller than using a builtin, so this should now be
preferred.