Smith: Remove pop instructions
The stack language is already built to consume the things it can
consume, and for identifiers, we don't want to consume them, but we can
index into the stack when we use one, so we don't always have to have
the one to use at the top.

So these pops only pollute the space of interesting programs with ones
that pointlessly add things to the stack and remove them again.

Remove the pop instructions entirely then. It doesn't even impact fuzzer
coverage much, which suggests they were not used so much in the corpus.
Which also makes sense, libFuzzer's REDUCE would have found smaller
programs that achieve the same thing without the pop.