Smith: Implement a rotate mutation
Instead of swap, probably it makes more sense to rotate the instructions
by 1. Or by a random amount. Maybe this is too hectic, and I should
limit it to just +/- 1? But let's see. My corpus is already pretty much
converged so it's hard to tell if this is helpful right now, but I like
the idea so I'm adding it.

This is where it would be useful to be able to see the mutation name,
not just "Custom" in libFuzzer's output. It _has_ to be possible. But
how?