Smith: Fuzz only programs that don't waste the stack
I saw many corpus entries that are smith programs where the output is
empty, but nonetheless the programs differ. Exploring that space is
wasteful. I wonder if an early exit, before we even run RCL, helps to
focus compute power more on exploring the interesting cases.

I don't have any hard measurements, but from me eyeballing the logs, it
seems to discover NEW and REDUCE cases at a slightly higher rate than
without this, even though the exec/s is not that much higher. Or is it?
It seems to go up over the runtime, both in the past and the current
fuzzer. Hmmmm.