Smith: Only keep samples that are minimal
I'm not sure if this is a good idea. On the one hand it keeps the corpus
small and full of interesting examples. But on the other hand, sometimes
we need multiple mutations to discover a new program. So I'm a bit
worried that this will get stuck in a local minimum. But maybe it will
also waste less CPU time on exploring pointless variations of a program?