Smith: Remove defaults on some underflows
On the one hand, this makes fewer inputs valid Smith programs, which
will make it harder for the fuzzer to find them. But on the other hand,
if these cases where the default helped were unlikely to lead to
interesting programs anyway, so maybe it's good to only unlock them when
the stack is properly filled, so that interesting interactions are
easier to discover in mutations?

It's all very hard to say, and very hard to measure without wasting a
huge amount of compute. But at least it makes the code simpler, so let's
go for that.

After this change, the fuzzer took a few minutes (~6 or so?) to return
to the previous level of coverage.