Smith: Remove identifier stack
I'm not sure if this is positive because it also enables more room for
syntax error when we expect something other than an identifier. But
maybe it simplifies Smith more than it harms and it speeds up the
fuzzer? I have one running for a few minutes now that is catching up to
one that is using the classic one, and it looks like it's catching up
very fast, but of course that is always the case at the start, so I'm
not sure. Maybe I should let it run over night and see which one gets
the furthest, coverage-wise?

Update: I let both of them run over night. The fuzzer without the
identifier stack started later (I don't remember how much, but maybe
half an hour, maybe an hour). Yet, it managed to achieve more coverage
(7805 vs. 7727, yesterday one with identifier stack was around 7500 and
the non-identifier at ~7200). So it looks like getting rid of the stack
is more helpful in speeding up the fuzzer, than harmful in generating
nonsense. Let's keep it.