Add golden test for multiline seq formatting
I thought this case where you have the [] on separate lines but the
inner seq is wide is not what I want, but after thinking some more,
I think it *is* what I want after all, the original logic makes sense.
The collection is wide or tall, but the elements can still individually
be wide.