Make chained expressions first-class in the CST
We do this to enable formatting a chain of field access in its entirety
as wide or tall. In the formatter, individual field expressions no
longer create groups, there is one group for the chain.

For now we do keep the tree structure (that in practice degenerates to
a singly linked list) in the AST, because it would be an invasive change
to make that a list too, and I'm not sure that is worth it. (It would
help somewhat to avoid stack overflow, but it may complicate the
typechecker and evaluator.) So the abstractor translates the list back
into a tree.