Format deep comprehensions tall, even if they fit
Deep comprehensions are probably complex comprehensions. When running
the autoformatter on some of the examples, I was not happy with how it
put some things too much on one line. So strike a balance here: if the
comprehension is deep (and therefore, I assume complex), format it as
tall even when it would have fit as wide.