Run new autoformatter on all examples
In the past I did not autoformat them because I was not happy with the
output yet, but now the formatter is good enough that I'm willing to
take this. There are still a few cases where I'm like ... meh, but that
requires human judgement that I think a formatter can never get right,
so I'm willing to stop arguing about this to get a formatter.

Possibly in the future I can add the magic trailing comma like Black has
in Python, to force collections to be tall. That could still be nice.
We can do that in a follow-up.