Handle multi-line holes in f-strings better
This is again a non-idempotency discovered by the fuzzer. Now an
expression in a hole can have blank lines or comments preceding it.

I am starting to see a pattern: all expressions (that can contain non-
code) need to be preceded by a separator. When they are preceded by
nothing (like was the case for the hole) or by a hard-coded space (like
was the case for the conditional), we may put a break or comment
directly after some content while it should go on its own line.

Are there more cases where an expression is not preceded by a separator?
I guess the fuzzer will find out!