Use expr_op for collection in for loops too
This change is similar to the one for conditionals. Like them, it stems
from the formatter printing a raw space before the collection, which is
problematic for non-code prefixes. Only here it did not surface as a non-
idempotency in the formatter because no non-code is allowed before the
collection. Still I think if you want to write this:

    [for x in let y = [1, 2, 3]; y: y * 2]

Then instead you should write this:

    [for x in (let y = [1, 2, 3]; y): y * 2]

It's much clearer and it creates fewer problems with formatting.