Fix formatter non-idempotency due to trailing comma
Respecting the trailing comma only for collections of >= 2 elements
previously masked this issue. One special case in the formatter is that
singleton list comprehensions don't get the trailing comma. But that
means that we should ignore it for the purpose of wide/tall.

This fixes the regression introduced in a070021. The problem was
discovered by the fuzzer.