Add Vega graphic as an example
Vega is a data format, though one could say configuration format, for
declaratively defining data visualizations. It can be generated from
Python using Altair, but how about generating it with RCL?

I don't know if this is a good idea, but it is at least possible.
And it's already way more convenient to write than writing json in
the online editor, because in RCL you don't have to quote the keys.