Add golden tests for "rcl build" command
I am a bit hesitant about making the tests write to the file system.
Maybe I should add a --dry-run mode to "rcl build" that instead prints
what files it would write to and what contents. But then we test less
of the real implementation, and there is something to say for testing
the real thing as much as possible, so let's do this for now. It may
cause trouble for the Nix flake checks, but outside of the Nix flake
it works.