Begin working on rationals and decimals
Let's finally add the float type. After some deliberation, I think I
want to represent them as decimals in scientific notation internally,
unless you do a division at which point they will turn into a rational.
I can turn the rational back into a decimal (which may be lossy) in
order to format it as a string; the rational itself is an implementation
detail, but it does allow us to get certain computations exactly right
where a float would not.