Add a Float type and Decimal value
Why the naming discrepancy? My thinking is that I want both Decimal and
Rational as internal representation, so Value can be decimal or rational.
But what is a good name for "decimal or rational"? I can think of
"number", but I want to reserve that for the supertype of int and this
one. "Rational" would be technically correct because decimals are also
rationals, just with a power of 10 denominator. But then what if I put
a sqrt function in there, or an approximation of pi? Sure technically
they are approximations, but I think this would _also_ be weird. So
let's go with float ... Also, this is what Cue calls them so there is
precedent.