Generalize inequalities from Int to Num in eval
This creates a new interesting problem though. The inequalities work
between Int and Float, but equality considers Int and Float different,
even when they are numerically the same. How to handle this in a sane
way? Probably it means that the distinction can only exist at the type
level, but not at runtime, and we should not have a runtime distinction
between Int and Float?