Pin cloudflare dns record example link
It appears that Cloudflare module for Terraform is experiencing and
API change in v5 [0]. The v5 is only in alpha release now, so I'm not
sure how long it will take to reach the release.

Pinning the link to 4.49.1 is a workaround for the time being.

For v5, the relevant at the moment page is [1].

0:
https://registry.terraform.io/providers/cloudflare/cloudflare/latest

1:
https://registry.terraform.io/providers/cloudflare/cloudflare/latest/docs/resources/dns_record