Make default help text briefer
Inspired by Cue, the `cue` commands without arguments prints somewhat
useful basic usage. I suppose `git` does so too. RCL used to print the
full reference help, which ends in the details for the color mode, which
is useful to have somewhere, but not that helpful as the first thing you
see when you run `rcl` (because the rest is too far up in your window).
So print the brief help if you run `rcl` without commands, and the full
help when requested with --help.

Document help text change in the changelog as well.