Parse all numbers into Value::Number, not Value::Int
This is one step in removing Value::Int. As part of this, move the
underscore stripping into Decimal::parse_str so we can avoid allocating
an intermediate string. (As if it matters with my inefficient number
parsing routine ... but hey, all bits help.)

This now breaks a *lot* of golden tests that use Int types, where type
inference now infers Number rather than Int, and also some places where
we check for Int but get a number. Perhaps I should do an intermediate
stage where the Int type is an alias for Number?