Phase out IntegerLit and Type::Int
There is now NumberLit, and we should be using Type::Number rather than
the Int type. This still causes a lot of fallout in the golden tests
which I still haven't updated, but at least by temporarily allowing Int
to be an alias for Number when parsing types, we can limit the fallout
a bit.