Remove Value::Int in favor of Value::Number
This again brings us one step closer to a uniform number type. It needs
a helper to extract an int when one is expected, but there were actually
very few places where that was needed. Maybe that is a good sign. But
maybe I can't think clearly because it's late and I should go to bed ...