Rename Int to Number in formatter golden tests
Again, the formatter doesn't care about type names, so these tests were
not failing before, but still, to make the tests a good example, and to
not cause confusion with outdated code that doesn't work, let's rename
things there too.

This is mostly a mechanical rename, but because the formatter cares
about length, in one case I had to change the example to make it still
exercise the same case.