Resolve to do in Decimal implementation
It doesn't apply, the comment was wrong. If there is an overflow, and
the signs are both positive, then m1 really is the greater one. I think
I didn't realize before that they are both positive? Anyway, nothing to
do to address this then.

I'm thinking if there is a test to add, but the logic was just wrong,
and there is already a test case that covers this control flow path, so
I think we're good.