Delete outdated idea files
These were helpful for initial exploration, but are confusing to have
lying around at this point. For example, they pollute the results when
I was 'git grep'ing for remaining references to Int.

The type system has long since been implemented, there is documentation
and a companion blog posts series about its design and implementation.
The idea directory is still there in the Git history for the future
historians, but it has no place any more in the default branch.