Correct bug when rounding doesn't change mantissa
This was discovered by inspecting the coverage report. The branch where
the mantissa didn't need adjustment was not yet covered, and then I
realized that returning the number itself was wrong. It was right in
the past, before I incorporated the exponent, but then I decided that
rounding should get rid of any exponent, and that broke this branch.
This also adds coverage for that branch.