Autoformat Python sources, make Mypy happy
The module name and path hack throws off Mypy, in different ways
depending on how you run it (outside Nix shell, inside Nix develop
shell, or as part of the flake check ...). Let's just forward-declare
it and sidestep problems.

Maybe I should make this script the source of truth and generate the
Pygments grammar from it. We could do that later, for now this works.