Add new builtin function std.format_json
For now this formats with unbounded width, and only json. Maybe I want
to give users control over the width, and also allow formatting to other
supported formats, perhaps using a more general 'std.format' function.
But json is probably the more useful one, and if you are putting things
into string literals, as long as I don't have output formats that
leverage  multi-line string literals (e.g. yaml that is not just json),
there is little point in formatting pretty with a width.

So this is experimental and subject to change, but let's try it out.
Documentation and tests will follow in a next commit, from a quick test
on the command line, it works.