Accept a new json-lines format in the CLI
I added a std.format_json function, and that required making the width
in the pretty-printer optional. With that in place, it is now easier to
support the JSON Lines format, something I wanted to do for some time
already.

This is a first step to accept the new format in the CLI. It does not
implement the formatting itself yet.