Add String.parse_number method
It doesn't work for all cases yet, for example it crashes on "abc".
Maybe it was not such a great idea to butcher the lexer for this, and we
should just put the error reporting in Decimal::from_str after all. And
support the numeric prefixes there too. But for now, this hooks things
up so there is _some_ behavior that mostly works.