Make golden for String.parse_number more powerful
If we test only for the overflow case, that admits an implementation
that just always returns overflow. If we also test the value just
before, it pins down the behavior much more.

I did realize because of this, that we can't parse i64::MIN, but I'm
fine with that for now. It doesn't work in RCL itself either, because
unary minus is an operator there.