Add indents and backet queries for Zed
The documentation for these is extremely sparse. I thought at firast
that this replaces the "brackets" section of the config.toml, especially
because that section is not documented, but in the upstream repo it does
still exist for the first-party plugins, so let's keep it here too. I
can experiment later and see what happens if we delete it.

First let's try and see if these new files do anything.