Tweak the bracket and indent queries for Zed
Bracket matching works nicely now. Indents mostly work, but there are
some bad cases. After a 'let' or opening bracket, Zed correctly indents,
but then when you type 'if ...:' in there, it suddenly dedents that. I
don't understand why, so I don't know how to fix it. But for now, this
is already better than the previous version of the plugin, which had
neither.