Add idea for 'rcl patch' command
When some data in a big configuration needs to come from other sources,
you can already dump it as json and then import it into RCL. But that
does spread your configuration over multiple files, and it makes Git
diffs harder to interpret, when context is missing.

I encountered a similar case at work a while back, where we had some
human-written config, but we wanted a script to edit it, while
preserving formatting and comments etc. Of course what you can always do
in that case is fall back to 'sed', and even though it might match false
positives, or create syntactically invalid files in theory, in practice
it works very well, especially if you add some marker comment.

But there *is* a proper way to do this. Parse into a concrete syntax
tree, and then do some surgery on the tree to splice in the new value.
And hey, what do you know, in RCL we happen to have a parser that builds
concrete syntax trees. And a pretty-printer for them. I thought at
first it would be difficult, because how do you navigate expressions?
But we can start with dicts that use record syntax, and including let
bindings is not too difficult either. Then we just walk the AST trying
to match identifier by identifier.