Make document name configurable in the loader
There are many things we can load from a string, we can give them better
names than just "input". This is in preparation for the 'rcl patch'
command, where I'll be loading more different documents at the same
time, and it will be more important to tell them apart.

This change can be made independently, so let's do it now to keep the
diff for the 'rcl patch' feature cleaner.