Implement first version of 'rcl patch'
This took me like 5 or 6 passes rewriting the return type of the tree
traversals, but in the end just Result and Option<Result> do the trick.

I was thinking about descending into other Seq elements, but maybe I
should not? If you have a loop, it makes no sense to have an AssocIdent
in there to patch. You might have a conditional, that *might* make
sense, but the things that you want to replace mechanically should be
simple configuration, probably not deeply nested logic. That logic may
be needed, but then just hoist the parameter that you want to adjust to
a higher level in the file, where it can be reached with a simpler path
expression.

There is one big issue remaining: the patched CST mixes spans from two
different documents, and the pretty-printer is *not* prepared to deal
with that. I could make it handle that, but it's kind of a weird
situation. Maybe alternatively, I could concat those two at the start,
and adjust the spans? Hmm, but then lines in errors would be wrong.
Probably we do need add multi-doc support in the CST pretty-printer ...