Enable the CST formatter to read from multiple docs
When we do surgery on the CST for the patch feature, and make it combine
inputs from multiple sources, the formatter needs to relove spans that
may not come from the same document.

I expected this to require horrible changes, and at first I was planning
to build a hash map of DocId to &str, but I already had this Inputs type
that does exactly what I need, and the changes are pretty minimal!