Tweak the handling of exponent 0 on integers
This makes the tests introduced in the parent commit pass. I say "tweak"
rather than "fix" because I think there is something to say for both
behaviors, and it's arguable whether you want a number with exponent but
without decimal point to be like a float or like an int, but at least
for consistency with 1e-1, 1e0, 1e1, I think it makes more sense to drop
the e0, than to turn the entire thing into a number with decimal point.