Unify output machinery between 'format' and 'patch'
If I have --in-place for formatting, it makes sense to have it for patch
as well. It's what you want to do anyway.

Patching multiple files at once makes less sense than formatting
multiple at once, but if we unify this stuff anyway, we might as well,
it keeps the code simpler.

One final thing not addressed here is that the output messages need to
be specialized to the command. TODO.