Specialize output for 'format' vs. 'patch'
These commands do something similar, and the way they report the result
is similar too, but not identical. Right now I do callbacks, but this
looks so messy! Maybe I should make an enum instead, and match on that.
Also, this is still incomplete, when printing the "would reformat", that
one needs to be parametrized too ...