Make printing format/patch stats a bit nicer
This took me a few attempt, first from callbacks, to returning an
unwieldy enum type, and then folding it in more and more, until I ended
up with this.

The type is a bit unholy, it can represent states that are invalid,
which we would eliminate with an enum, but then the pattern matches
would all turn into big enum-style matches, and the code would be 3x
as long, which makes it ugly and harder to understand. There is
something to say for terseness in this case.