Flatten the Seq construct in the CST
I did this in the past for expressions, but for Seq inside collection
literals, they were still in this tree shape. There are many places
where we then need to descend into the tree to find the entry at the
lowest level. For the formatter at least it's nicer to everything but
the end of the chain as a single construct, so we can make the entire
thing wide or tall, that's been on my to do list for a long time. But
what is really triggering this now, is the 'rcl patch' command, where
the code for traversal could be a bit simpler when the list is flat.

After an initial step of only flattening Stmt into Statements, that
turned out to be insufficient. We need to split out Yield instead, just
like we already have in the AST, and then flatten everything that
precedes that yield.