Update the parser for the new cst::Seq
This uncovered one problem in the formatter, which is that a non-code
prefix will now make the entire Seq tall, so if you have something like

    // Double all the xs.
    for x in xs: x * 2

that's a single group, forced tall, so it becomes

    // Double all the xs.
    for x in xs:
    x * 2

Which is not what I intended, the first prefix should be hoisted out of
the Seq group.

Aside from that, this now handles comments in more places, which is a
sign that the new CST is an improvement. I'll add tests for that in
follow-up commits. There are even more places where we could consume
rather than prohibit non-code, as long as we properly return it. We can
look at that as a follow-up as well.

Overall, this feels like a nice improvement. We can even get rid of
parse_prefixed, which somehow feels like the right thing?