Ensure that comments don't force tall seq elements
This restores the behavior before the refactor, where this was handled
through collections storing a List<Prefixed<Seq>> rather than List<Seq>.

There was no test to detect the previous suboptimal format, so this adds
one as well.

At first I had a premature optimization here to skip the concat! node
if the prefix was empty, but it was harmful for test coverage because
an empty non-code returns Doc::Concat([]) rather than Doc::Empty, and
I forgot that concat! is based on the + operator, which already
optimizes all these cases. The concat expands to:

    let mut result = Doc::Empty;
    result = result + Doc::Concat(prefix);
    result = result + Doc::Group(parts);

After the first +, it's just the Concat([]) due to the rule that Empty
can be omitted. After the second +, it's just the group, due to the rule
that Concat([]) can be omitted. So, I was foolish to complicate the code
to try and do better. Maybe there is still a small runtime hit, but so
far formatting is pretty much instant anyway, so nothing to worry about.