Make expression parser not force non-code after
Sometimes the parser needs to peek ahead, but because of non-code, that
means we need to consume it, and then we can't allow a comment there.
Instead, we can peek through non-code, and still report an error later
if we did go down a branch and there was a comment, but then at least we
don't eat the comment and non-code if we don't go down that branch, and
that means that we can for example leave the trailing comment in a
collection.

I wonder what the performance impact on the parser is. I'm not measuring
at this point.