Add tests for handling of split suffixes
This was the one regression in test coverage for the parser. While
adding this test, I realized we need to also make sure to not create
consecutive whitespace, as it would create a non-idempotency in the
formatter. The fuzzer has not discovered that yet, though I suspect it
would after a while.