Make matching in 'rcl patch' more principled
The preceding commits in this series were all done against the previous
iteration of the CST, where Seq was nested. I now completed a refactor
where every Seq contains a flat list of the preceding "control items",
and an inner body that is directly accessible.

Now that this 'rcl patch' series is rebased on top of the CST rework,
we can start to leverage the new structure, and indeed the matching
becomes a lot simpler to do exhaustively.